// src/app/shared/models/user.ts
import { JobParameters } from "./JobParameters";

export class User {
  UserId: number;
  UserName: string;
  UserTel: string;
  UserMail: string;
  UserCityId: number;
  UserSubId: number;
  UserPartId: number;
  UserIsChizuk: boolean;
  UserIsUpdate: boolean;
  password: string;
  cvURL: string;
  UserIsSmartAgent: boolean;
  UserSmartAgentTime: string;
  // jobParameters:JobParameters;
  //     constructor(userId,userName,userTel,userMail,userCityId,userSubId,userPartId,userIsChizuk,userIsUpdate){
  //   this.UserId  =userId;
  //   this.UserName=userName;
  //   this.UserMail=userMail;
  //   this.UserCityId=userCityId;
  //   this.UserSubId =userSubId;
  //   this.UserPartId=userPartId;
  //   this.UserIsChizuk=userIsChizuk;
  //   this.UserIsUpdate=userIsUpdate;
  //     }
}